requirements.txt
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.0.76
pytesseract==0.3.10
pyautogui==0.9.54
pillow==10.0.0
pywin32==306
matplotlib==3.7.2
mss==9.0.1
screeninfo==0.8.1
requests==2.31.0
configparser==6.0.0